CDRIVER-752 int overflow in mongoc_cond_timedwait
On Solaris, at least, timeval.tv_sec is 32 bits. Cast to int64_t before
multiplying by 1000 to avoid overflow when converting from seconds to ms.

+++ src/mongoc/mongoc-thread-private.h
@@ -45,7 +45,7 @@ mongoc_cond_timedwait (pthread_cond_t  *cond,
 
    bson_gettimeofday (&tv);
 
-   msec = (tv.tv_sec * 1000) + (tv.tv_usec / 1000) + timeout_msec;
+   msec = ((int64_t)tv.tv_sec * 1000) + (tv.tv_usec / 1000) + timeout_msec;
 
    to.tv_sec = msec / 1000;
    to.tv_nsec = (msec % 1000) * 1000 * 1000;
